// services/API-service/src/api/admin-area-dynamic-data/enum/dynamic-indicator.ts
export enum DynamicIndicator {
  populationAffected = 'population_affected',
  populationAffectedPercentage = 'population_affected_percentage',
  potentialCasesU9 = 'potential_cases_U9',
  potentialCasesU5 = 'potential_cases_U5',
  potentialCases65 = 'potential_cases_65',
  potentialCases = 'potential_cases',
  potentialThreshold = 'potential_cases_threshold',
  alertThreshold = 'alert_threshold',
}